backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-groq
langchain-community
langchain-text-splitters
langchain-huggingface
pydantic-settings
python-dotenv
pydantic
duckduckgo-search
numexpr
faiss-cpu
pypdf
python-multipart
